requirements.txt
numpy ==1.13
torch==1.5.0
scipy
PyYaml 
matplotlib
